Had to fix small error, updated package
From now on, I will only tell you the new package version.

Current version
2.0.3

+++ src/package_pyWrskp/src/pyWrskp/odds_and_ends.py
@@ -23,7 +23,7 @@ def number_input(question,
 
 
 def loading(time):
-    loops = time * 1.5
+    loops = int(time * 1.5)
     for i in range(loops):
         print('|', end="\r")
         sleep(0.5)
